start in normal mapping

+++ includes/rt.h
@@ -6,7 +6,7 @@
 /*   By: lwyl-the <lwyl-the@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/11 19:14:44 by rgyles            #+#    #+#             */
-/*   Updated: 2019/04/11 20:06:21 by lwyl-the         ###   ########.fr       */
+/*   Updated: 2019/04/12 17:13:25 by lwyl-the         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,6 @@ typedef struct		s_rt
 	double			sample_step;
 	double			sample_center;
 	double			t_closest;
-	//int				*light_array;
 
 	t_vec3			angle;
 	t_vec3			camera;
@@ -90,7 +89,7 @@ double				mix(double dist_a, double dist_b);
 int					get_color(t_vec3 *dir, t_shape *shape, t_rt *rt, int depth);
 double				get_light(t_vec3 *dir, t_shape *shape, t_rt *rt);
 void				get_normal(t_shape *shape);
-int					shadow(t_vec3 *orig, t_vec3 dir, t_shape *head_shapes, double max_distance);
+double					shadow(t_vec3 *orig, t_vec3 dir, t_shape *head_shapes, double max_distance);
 int					reflection(t_vec3 *dir, t_shape *shape, t_rt *rt, int depth);
 int					emission(t_shape *shape, t_rt *rt, int depth);
 //int					trace_ray(t_coord *ray, t_rt *rt); //int depth);
